+++ playwright/e2e/read-receipts/high-level.spec.ts
@@ -19,77 +19,6 @@ limitations under the License.
 import { customEvent, many, test } from ".";
 
 test.describe("Read receipts", () => {
-    test.describe("Message ordering", () => {
-        test.describe("in the main timeline", () => {
-            test.fixme(
-                "A receipt for the last event in sync order (even with wrong ts) marks a room as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for a non-last event in sync order (even when ts makes it last) leaves room unread",
-                () => {},
-            );
-        });
-
-        test.describe("in threads", () => {
-            // These don't pass yet - we need MSC4033 - we don't even know the Sync order yet
-            test.fixme(
-                "A receipt for the last event in sync order (even with wrong ts) marks a thread as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for a non-last event in sync order (even when ts makes it last) leaves thread unread",
-                () => {},
-            );
-
-            // These pass now and should not later - we should use order from MSC4033 instead of ts
-            // These are broken out
-            test.fixme(
-                "A receipt for last threaded event in ts order (even when it was received non-last) marks a thread as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for non-last threaded event in ts order (even when it was received last) leaves thread unread",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for last threaded edit in ts order (even when it was received non-last) marks a thread as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for non-last threaded edit in ts order (even when it was received last) leaves thread unread",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for last threaded reaction in ts order (even when it was received non-last) marks a thread as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for non-last threaded reaction in ts order (even when it was received last) leaves thread unread",
-                () => {},
-            );
-        });
-
-        test.describe("thread roots", () => {
-            test.fixme(
-                "A receipt for last reaction to thread root in sync order (even when ts makes it last) marks room as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for non-last reaction to thread root in sync order (even when ts makes it last) leaves room unread",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for last edit to thread root in sync order (even when ts makes it last) marks room as read",
-                () => {},
-            );
-            test.fixme(
-                "A receipt for non-last edit to thread root in sync order (even when ts makes it last) leaves room unread",
-                () => {},
-            );
-        });
-    });
-
     test.describe("Ignored events", () => {
         test("If all events after receipt are unimportant, the room is read", async ({
             roomAlpha: room1,
@@ -412,81 +341,6 @@ test.describe("Read receipts", () => {
             await util.assertReadThread("Root1");
             await util.assertReadThread("Root2");
             await util.assertReadThread("Root3");
-        });
-    });
-
-    test.describe("Room list order", () => {
-        test("Rooms with unread messages appear at the top of room list if 'unread first' is selected", async ({
-            roomAlpha: room1,
-            roomBeta: room2,
-            util,
-            msg,
-            page,
-        }) => {
-            await util.goTo(room2);
-
-            // Display the unread first room
-            await util.toggleRoomUnreadOrder();
-            await util.receiveMessages(room1, ["Msg1"]);
-            await page.reload();
-
-            // Room 1 has an unread message and should be displayed first
-            await util.assertRoomListOrder([room1, room2]);
-        });
-
-        test("Rooms with unread threads appear at the top of room list if 'unread first' is selected", async ({
-            roomAlpha: room1,
-            roomBeta: room2,
-            util,
-            msg,
-        }) => {
-            await util.goTo(room2);
-            await util.receiveMessages(room1, ["Msg1"]);
-            await util.markAsRead(room1);
-            await util.assertRead(room1);
-
-            // Display the unread first room
-            await util.toggleRoomUnreadOrder();
-            await util.receiveMessages(room1, [msg.threadedOff("Msg1", "Resp1")]);
-            await util.saveAndReload();
-
-            // Room 1 has an unread message and should be displayed first
-            await util.assertRoomListOrder([room1, room2]);
-        });
-    });
-
-    test.describe("Notifications", () => {
-        test.describe("in the main timeline", () => {
-            test.fixme("A new message that mentions me shows a notification", () => {});
-            test.fixme(
-                "Reading a notifying message reduces the notification count in the room list, space and tab",
-                () => {},
-            );
-            test.fixme(
-                "Reading the last notifying message removes the notification marker from room list, space and tab",
-                () => {},
-            );
-            test.fixme("Editing a message to mentions me shows a notification", () => {});
-            test.fixme("Reading the last notifying edited message removes the notification marker", () => {});
-            test.fixme("Redacting a notifying message removes the notification marker", () => {});
-        });
-
-        test.describe("in threads", () => {
-            test.fixme("A new threaded message that mentions me shows a notification", () => {});
-            test.fixme("Reading a notifying threaded message removes the notification count", () => {});
-            test.fixme(
-                "Notification count remains steady when reading threads that contain seen notifications",
-                () => {},
-            );
-            test.fixme(
-                "Notification count remains steady when paging up thread view even when threads contain seen notifications",
-                () => {},
-            );
-            test.fixme(
-                "Notification count remains steady when paging up thread view after mark as unread even if older threads contain notifications",
-                () => {},
-            );
-            test.fixme("Redacting a notifying threaded message removes the notification marker", () => {});
         });
     });
 });
