+++ playwright/e2e/read-receipts/new-messages-in-threads.spec.ts
@@ -20,151 +20,6 @@ import { many, test } from ".";
 
 test.describe("Read receipts", () => {
     test.describe("new messages", () => {
-        test.describe("in the main timeline", () => {
-            test("Receiving a message makes a room unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I am in a different room
-                await util.goTo(room1);
-                await util.assertRead(room2);
-
-                // When I receive some messages
-                await util.receiveMessages(room2, ["Msg1"]);
-
-                // Then the room is marked as unread
-                await util.assertUnread(room2, 1);
-            });
-            test("Reading latest message makes the room read", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I have some unread messages
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-
-                // When I read the main timeline
-                await util.goTo(room2);
-
-                // Then the room becomes read
-                await util.assertRead(room2);
-            });
-            test("Reading an older message leaves the room unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given there are lots of messages in a room
-                await util.goTo(room1);
-                await util.receiveMessages(room2, many("Msg", 30));
-                await util.assertUnread(room2, 30);
-
-                // When I jump to one of the older messages
-                await msg.jumpTo(room2.name, "Msg0001");
-
-                // Then the room is still unread, but some messages were read
-                await util.assertUnreadLessThan(room2, 30);
-            });
-            test("Marking a room as read makes it read", async ({ roomAlpha: room1, roomBeta: room2, util, msg }) => {
-                // Given I have some unread messages
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-
-                // When I mark the room as read
-                await util.markAsRead(room2);
-
-                // Then it is read
-                await util.assertRead(room2);
-            });
-            test("Receiving a new message after marking as read makes it unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I have marked my messages as read
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-
-                // When I receive a new message
-                await util.receiveMessages(room2, ["Msg2"]);
-
-                // Then the room is unread
-                await util.assertUnread(room2, 1);
-            });
-            test("A room with a new message is still unread after restart", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I have an unread message
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-
-                // When I restart
-                await util.saveAndReload();
-
-                // Then I still have an unread message
-                await util.assertUnread(room2, 1);
-            });
-            test("A room where all messages are read is still read after restart", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I have read all messages
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-                await util.goTo(room2);
-                await util.assertRead(room2);
-
-                // When I restart
-                await util.saveAndReload();
-
-                // Then all messages are still read
-                await util.assertRead(room2);
-            });
-            test("A room that was marked as read is still read after restart", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given I have marked all messages as read
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1"]);
-                await util.assertUnread(room2, 1);
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-
-                // When I restart
-                await util.saveAndReload();
-
-                // Then all messages are still read
-                await util.assertRead(room2);
-            });
-        });
-
         test.describe("in threads", () => {
             test("Receiving a message makes a room unread", async ({
                 roomAlpha: room1,
@@ -448,101 +303,6 @@ test.describe("Read receipts", () => {
                 await util.assertRead(room2);
                 await util.goTo(room2);
                 await util.assertReadThread("Msg1");
-            });
-        });
-
-        test.describe("thread roots", () => {
-            test("Reading a thread root does not mark the thread as read", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given a thread exists
-                await util.goTo(room1);
-                await util.receiveMessages(room2, ["Msg1", msg.threadedOff("Msg1", "Resp1")]);
-                await util.assertUnread(room2, 1); // (Sanity)
-
-                // When I read the main timeline
-                await util.goTo(room2);
-
-                // Then room doesn't appear unread but the thread does
-                await util.assertRead(room2);
-                await util.assertUnreadThread("Msg1");
-            });
-
-            test("Reading a thread root within the thread view marks it as read in the main timeline", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given lots of messages are on the main timeline, and one has a thread off it
-                await util.goTo(room1);
-                await util.receiveMessages(room2, [
-                    ...many("beforeThread", 30),
-                    "ThreadRoot",
-                    msg.threadedOff("ThreadRoot", "InThread"),
-                    ...many("afterThread", 30),
-                ]);
-                await util.assertUnread(room2, 61); // Sanity
-
-                // When I jump to an old message and read the thread
-                await msg.jumpTo(room2.name, "beforeThread0000");
-                // When the thread is opened, the timeline is scrolled until the thread root reached the center
-                await util.openThread("ThreadRoot");
-
-                // Then the thread root is marked as read in the main timeline,
-                // 30 remaining messages are unread - 7 messages are displayed under the thread root
-                await util.assertUnread(room2, 30 - 7);
-            });
-
-            test("Creating a new thread based on a reply makes the room unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given a message and reply exist and are read
-                await util.goTo(room1);
-                await util.receiveMessages(room2, ["Msg1", msg.replyTo("Msg1", "Reply1")]);
-                await util.goTo(room2);
-                await util.assertRead(room2);
-                await util.goTo(room1);
-                await util.assertRead(room2);
-
-                // When I receive a thread message created on the reply
-                await util.receiveMessages(room2, [msg.threadedOff("Reply1", "Resp1")]);
-
-                // Then the thread is unread
-                await util.goTo(room2);
-                await util.assertUnreadThread("Reply1");
-            });
-
-            test("Reading a thread whose root is a reply makes the thread read", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given an unread thread off a reply exists
-                await util.goTo(room1);
-                await util.receiveMessages(room2, [
-                    "Msg1",
-                    msg.replyTo("Msg1", "Reply1"),
-                    msg.threadedOff("Reply1", "Resp1"),
-                ]);
-                await util.assertUnread(room2, 2);
-                await util.goTo(room2);
-                await util.assertRead(room2);
-                await util.assertUnreadThread("Reply1");
-
-                // When I read the thread
-                await util.openThread("Reply1");
-
-                // Then the room and thread are read
-                await util.assertRead(room2);
-                await util.assertReadThread("Reply1");
             });
         });
     });
