+++ playwright/e2e/read-receipts/reactions-in-threads.spec.ts
@@ -20,82 +20,6 @@ import { test, expect } from ".";
 
 test.describe("Read receipts", () => {
     test.describe("reactions", () => {
-        test.describe("in the main timeline", () => {
-            test("Receiving a reaction to a message does not make a room unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", "Msg2"]);
-                await util.assertUnread(room2, 2);
-
-                // When I read the main timeline
-                await util.goTo(room2);
-                await util.assertRead(room2);
-
-                await util.goTo(room1);
-                await util.receiveMessages(room2, [msg.reactionTo("Msg2", "🪿")]);
-                await util.assertRead(room2);
-            });
-            test("Reacting to a message after marking as read does not make the room unread", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", "Msg2"]);
-                await util.assertUnread(room2, 2);
-
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-
-                await util.receiveMessages(room2, [msg.reactionTo("Msg2", "🪿")]);
-                await util.assertRead(room2);
-            });
-            test("A room with an unread reaction is still read after restart", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", "Msg2"]);
-                await util.assertUnread(room2, 2);
-
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-
-                await util.receiveMessages(room2, [msg.reactionTo("Msg2", "🪿")]);
-                await util.assertRead(room2);
-
-                await util.saveAndReload();
-                await util.assertRead(room2);
-            });
-            test("A room where all reactions are read is still read after restart", async ({
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", "Msg2", msg.reactionTo("Msg2", "🪿")]);
-                await util.assertUnread(room2, 2);
-
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-
-                await util.saveAndReload();
-                await util.assertRead(room2);
-            });
-        });
-
         test.describe("in threads", () => {
             test("A reaction to a threaded message does not make the room unread", async ({
                 roomAlpha: room1,
@@ -279,98 +203,6 @@ test.describe("Read receipts", () => {
                 await page.locator(".mx_EmojiPicker_body").getByText("😀").first().click();
                 await page.locator(".mx_ThreadPanel").getByLabel("Mae reacted with 😀").click();
                 await expect(await page.locator(".mx_ThreadPanel").getByLabel("Mae reacted with 😀")).not.toBeVisible();
-            });
-        });
-
-        test.describe("thread roots", () => {
-            test("A reaction to a thread root does not make the room unread", async ({
-                page,
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given a read thread root exists
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", msg.threadedOff("Msg1", "Reply1")]);
-                await util.assertUnread(room2, 1);
-                await util.goTo(room2);
-                await util.openThread("Msg1");
-                await util.assertRead(room2);
-                await util.assertReadThread("Msg1");
-
-                // When someone reacts to it
-                await util.goTo(room1);
-                await util.receiveMessages(room2, [msg.reactionTo("Msg1", "🪿")]);
-                await page.waitForTimeout(200);
-
-                // Then the room is still read
-                await util.assertRead(room2);
-                // as is the thread
-                await util.assertReadThread("Msg1");
-            });
-
-            test("Reading a reaction to a thread root leaves the room read", async ({
-                page,
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given a read thread root exists
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", msg.threadedOff("Msg1", "Reply1")]);
-                await util.assertUnread(room2, 1);
-                await util.goTo(room2);
-                await util.openThread("Msg1");
-                await util.assertRead(room2);
-
-                // And the reaction to it does not make us unread
-                await util.goTo(room1);
-                await util.receiveMessages(room2, [msg.reactionTo("Msg1", "🪿")]);
-                await util.assertRead(room2);
-                await util.assertReadThread("Msg1");
-
-                // When we read the reaction and go away again
-                await util.goTo(room2);
-                await util.openThread("Msg1");
-                await util.assertRead(room2);
-                await util.goTo(room1);
-                await page.waitForTimeout(200);
-
-                // Then the room is still read
-                await util.assertRead(room2);
-                await util.assertReadThread("Msg1");
-            });
-
-            test("Reacting to a thread root after marking as read makes the room unread but not the thread", async ({
-                page,
-                roomAlpha: room1,
-                roomBeta: room2,
-                util,
-                msg,
-            }) => {
-                // Given a thread root exists
-                await util.goTo(room1);
-                await util.assertRead(room2);
-                await util.receiveMessages(room2, ["Msg1", msg.threadedOff("Msg1", "Reply1")]);
-                await util.assertUnread(room2, 1);
-
-                // And we have marked the room as read
-                await util.markAsRead(room2);
-                await util.assertRead(room2);
-                await util.assertReadThread("Msg1");
-
-                // When someone reacts to it
-                await util.receiveMessages(room2, [msg.reactionTo("Msg1", "🪿")]);
-                await page.waitForTimeout(200);
-
-                // Then the room is still read
-                await util.assertRead(room2);
-                // as is the thread
-                await util.assertReadThread("Msg1");
             });
         });
     });
