+++ playwright/e2e/settings/appearance-user-settings-tab/appearance-user-settings-tab.spec.ts
@@ -60,113 +60,4 @@ test.describe("Appearance user settings tab", () => {
         // Assert that the font-family value was removed
         await expect(page.locator("body")).toHaveCSS("font-family", '""');
     });
-
-    test.describe("Message Layout Panel", () => {
-        test.beforeEach(async ({ app, user, util }) => {
-            await util.createAndDisplayRoom();
-            await util.assertModernLayout();
-            await util.openAppearanceTab();
-        });
-
-        test("should change the message layout from modern to bubble", async ({ page, app, user, util }) => {
-            await util.assertScreenshot(util.getMessageLayoutPanel(), "message-layout-panel-modern.png");
-
-            await util.getBubbleLayout().click();
-
-            // Assert that modern are irc layout are not selected
-            await expect(util.getBubbleLayout()).toBeChecked();
-            await expect(util.getModernLayout()).not.toBeChecked();
-            await expect(util.getIRCLayout()).not.toBeChecked();
-
-            // Assert that the room layout is set to bubble layout
-            await util.assertBubbleLayout();
-            await util.assertScreenshot(util.getMessageLayoutPanel(), "message-layout-panel-bubble.png");
-        });
-
-        test("should enable compact layout when the modern layout is selected", async ({ page, app, user, util }) => {
-            await expect(util.getCompactLayoutCheckbox()).not.toBeChecked();
-
-            await util.getCompactLayoutCheckbox().click();
-            await util.assertCompactLayout();
-        });
-
-        test("should disable compact layout when the modern layout is not selected", async ({
-            page,
-            app,
-            user,
-            util,
-        }) => {
-            await expect(util.getCompactLayoutCheckbox()).not.toBeDisabled();
-
-            // Select the bubble layout, which should disable the compact layout checkbox
-            await util.getBubbleLayout().click();
-            await expect(util.getCompactLayoutCheckbox()).toBeDisabled();
-        });
-    });
-
-    test.describe("Theme Choice Panel", () => {
-        test.beforeEach(async ({ app, user, util }) => {
-            // Disable the default theme for consistency in case ThemeWatcher automatically chooses it
-            await util.disableSystemTheme();
-            await util.openAppearanceTab();
-        });
-
-        test("should be rendered with the light theme selected", async ({ page, app, util }) => {
-            // Assert that 'Match system theme' is not checked
-            await expect(util.getMatchSystemThemeCheckbox()).not.toBeChecked();
-
-            // Assert that the light theme is selected
-            await expect(util.getLightTheme()).toBeChecked();
-            // Assert that the dark and high contrast themes are not selected
-            await expect(util.getDarkTheme()).not.toBeChecked();
-            await expect(util.getHighContrastTheme()).not.toBeChecked();
-
-            await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-light.png");
-        });
-
-        test("should disable the themes when the system theme is clicked", async ({ page, app, util }) => {
-            await util.getMatchSystemThemeCheckbox().click();
-
-            // Assert that the themes are disabled
-            await expect(util.getLightTheme()).toBeDisabled();
-            await expect(util.getDarkTheme()).toBeDisabled();
-            await expect(util.getHighContrastTheme()).toBeDisabled();
-
-            await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-match-system-enabled.png");
-        });
-
-        test("should change the theme to dark", async ({ page, app, util }) => {
-            // Assert that the light theme is selected
-            await expect(util.getLightTheme()).toBeChecked();
-
-            await util.getDarkTheme().click();
-
-            // Assert that the light and high contrast themes are not selected
-            await expect(util.getLightTheme()).not.toBeChecked();
-            await expect(util.getDarkTheme()).toBeChecked();
-            await expect(util.getHighContrastTheme()).not.toBeChecked();
-
-            await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-dark.png");
-        });
-
-        test.describe("custom theme", () => {
-            test.use({
-                labsFlags: ["feature_custom_themes"],
-            });
-
-            test("should render the custom theme section", async ({ page, app, util }) => {
-                await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-custom-theme.png");
-            });
-
-            test("should be able to add and remove a custom theme", async ({ page, app, util }) => {
-                await util.addCustomTheme();
-
-                await expect(util.getCustomTheme()).not.toBeChecked();
-                await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-custom-theme-added.png");
-
-                await util.removeCustomTheme();
-                await expect(util.getThemePanel()).toMatchScreenshot("theme-panel-custom-theme.png");
-            });
-        });
-    });
 });
